Exercises 2.6-2.10
Creating a phonebook website that allows a user to input a new name and number, prevent duplicates and filter our list of names. This can probably be consolidated but I made 5 components when I really could get away with 4 maybe? Handlers could probably be in their own file as well but this is fine.